Add files via upload
Geometry is one of the important sections in mathematics. 

This program is coded to conduct Quiz’ related to geometrical aspects like that of Circles, Squares and Planes.
The program enables the user to Take Quiz, View all user records and to view user records by help of key words like name, class, section and roll number. The program also provides ADMIN controls like that of deletion of record and a password system for security purpose.

This program was coded using C++ and can be utilized in school geometry assessment tests in the form of quiz to assess 
the student's performance in a class.